# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Company IT Service Platform API
  version: 0.1.0
  description: REST API for tickets, users, assets, and support copilot analysis.
servers:
  - url: http://localhost:4000
paths:
...
  /api/tickets/{id}:
    get:
      summary: Get ticket by id
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Ticket details
        "404":
          description: Ticket not found
  /api/tickets/{id}/status:
    patch:
      summary: Update ticket status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated ticket
...